+++ scopeql/CHANGELOG.md
@@ -7,10 +7,11 @@ All significant changes to this software be documented in this file.
 ### Breaking Changes
 
 * `scopeql -c < script.sql` no longer supported, use `scopeql run -f script.sql` instead.
-* `scopeql -c -` no longer supported, use `scopeql run [ -f <filename> ] [ -c <statement> ]` instead.
+* `scopeql -c -` no longer supported.
 
 ### New Features
 
+* Support `scopeql run <statement>` to run statement directly.
 * Support `scopeql run -f <filename>` to run script from file.
 
 ## v0.2.2 (2025-12-08)